% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AlphaMissenseSample}
\alias{AlphaMissenseSample}
\title{AlphaMissense Sample Dataset}
\format{
A data table with 6 variables:
\describe{
  \item{CHROM}{Chromosome of the variant (with 'chr' prefix).}
  \item{POS}{Position of the variant on the chromosome.}
  \item{REF}{Reference allele.}
  \item{ALT}{Alternate allele.}
  \item{genome}{Genome build, here 'hg38'.}
  \item{uniprot_id}{UniProt identifier for the protein.}
  \item{transcript_id}{Transcript identifier.}
  \item{protein_variant}{Amino acid change resulting from the variant.}
  \item{am_pathogenicity}{Pathogenicity score for the variant.}
  \item{am_class}{Classification of the variant, e.g., 'likely_benign'.}
}
}
\source{
Jun Cheng et al., "Accurate proteome-wide missense variant effect prediction
with AlphaMissense." Science 381, eadg7492 (2023). DOI: 10.1126/science.adg7492.
The larger dataset includes 22.8 million variants classified as likely pathogenic
and 40.9 million as likely benign.
}
\usage{
AlphaMissenseSample
}
\description{
A sample dataset containing missense variant predictions from the AlphaMissense
community dataset resource. This dataset is a small extract from a larger
collection of 71 million missense variant predictions that saturate the human proteome.
Each variant is characterized by a single nucleotide change resulting in an altered
amino acid. This sample includes information on the chromosomal location, reference and
alternate alleles, genome build, protein identifiers, and the predicted pathogenicity
and class of the variants.
}
\examples{
data("AlphaMissenseSample")
head(AlphaMissenseSample)
}
\keyword{datasets}
