% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classVis.R
\name{classVis}
\alias{classVis}
\title{Visualize Proportions of Pathogenicity Classes}
\usage{
classVis(data, groups = NULL)
}
\arguments{
\item{data}{A data table containing at least 'group', 'am_class' columns.}

\item{groups}{A character vector specifying the groups to include in the plot.
If NULL, all groups in 'data' are used.}
}
\value{
A ggplot object representing the stacked barplot of pathogenicity class distribution.
}
\description{
This function creates a stacked barplot to visualize the proportions
of each pathogenicity class within and between specified groups.
}
\examples{
\dontrun{
  classVis(predScoreSample, groups = c("0h", "1h"))
}
}
