% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF.R
\name{createVCFDataFrame}
\alias{createVCFDataFrame}
\title{Create DataFrame for VCF Files}
\usage{
createVCFDataFrame(base_dir, vcf_pattern, group_pattern, sample_pattern)
}
\arguments{
\item{base_dir}{Directory containing VCF files.}

\item{vcf_pattern}{Regular expression to match VCF files.}

\item{group_pattern}{Regular expression to extract group names from file paths.}

\item{sample_pattern}{Regular expression to extract sample names from file paths.}
}
\value{
DataFrame with columns: group, sample, file_name.
}
\description{
This function scans a directory for VCF files matching a given pattern and
creates a DataFrame with group, sample, and file path information.
}
\examples{
vcf_files_path <- system.file("extdata", package = "MissensePathoR")
vcf_df <- createVCFDataFrame(vcf_files_path,
                             "*.vcf",
                             "([0-9]+)h",
                             "Rep([0-9]+)")

}
