% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffSNPs.R
\name{diffSNPs}
\alias{diffSNPs}
\title{Differential SNP Analysis}
\usage{
diffSNPs(data, groupControl)
}
\arguments{
\item{data}{A data frame containing 'uniprot_id', 'sample_name', and 'am_class'.}

\item{groupControl}{The reference level for the 'group' factor in the model design.}
}
\value{
A table of differentially expressed SNPs.
}
\description{
This function performs differential SNP analysis based on pathogenicity scores and variant frequencies.
We used variant frequency as the pseudo gene count and normalized it using CPM (Counts Per Million).
Incorporates pathogenicity score for weighted CPM calculation and uses a Quasi-likelihood model to account for over-dispersion.
}
\examples{
\dontrun{
  result <- mapGene(predScoreSample)
  result$sample_name = paste0(result$sample_name,"_",result$group)
  diffOut <- diffSNPs(result, "0h")
  head(diffOut$table)
}
}
