% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichSNP.R
\name{enrichSNP}
\alias{enrichSNP}
\title{Pathway Enrichment using Thresholded Over-representation Analysis for SNPs}
\usage{
enrichSNP(diffOut)
}
\arguments{
\item{diffOut}{A TopTags object containing differential expression results with columns 'FDR' and 'logFC'.}
}
\value{
A list containing the enrichment results for up-regulated, down-regulated, and thresholded genes.
}
\description{
Performs gene set enrichment analysis on identified up- and down-regulated genes using the gprofiler2 package.
This function applies a significance threshold (FDR <= 0.05) and uses log fold change (logFC) to categorize genes into up-regulated and down-regulated groups.
In edge cases where fewer than one gene is identified as significantly differentially expressed, the function adjusts the gene lists accordingly.
For instance, if no genes meet the significance criteria, a warning is issued. If only a few genes are identified, the function extends the gene list based on fold change to meet the minimum threshold for enrichment analysis.
}
\examples{
\dontrun{
  enrichOut <- enrichSNP(diffOut)
}
}
