% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayViz.R
\name{pathwayViz}
\alias{pathwayViz}
\title{Visualize Enrichment Pathways}
\usage{
pathwayViz(enrichOut, type = "sig")
}
\arguments{
\item{enrichOut}{A list containing enrichment results as obtained from
an enrichment analysis function like \code{gost}. This list should contain
elements named 'sig', 'up', and 'down', each with a sub-element 'go_bp'.}

\item{type}{Character string specifying the type of genes to visualize.
Should be one of 'sig' (significant), 'up' (up-regulated), or 'down' (down-regulated).
Default is 'sig'.}
}
\value{
A ggplot object representing the top 10 enriched pathways for the
  specified gene set.
}
\description{
Provides dot plots displaying enriched pathways for a given set of genes,
enhancing the interpretability of pathway analysis results.
}
\examples{
\dontrun{
  pathwayViz(enrichOut, type = "sig")
}
}
