% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichSNP.R
\name{performGost}
\alias{performGost}
\title{Helper function, Perform GOST Analysis}
\usage{
performGost(gene_list, query_type)
}
\arguments{
\item{gene_list}{A vector of gene identifiers.}

\item{query_type}{A character string specifying the type of query.}
}
\value{
A list containing the GOST result, a plot of the GOST result, and GO Biological Process results.
}
\description{
This helper function uses the g:Profiler tool for functional mapping and performs a GOST analysis on the provided gene list.
}
