% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predScoreSample}
\alias{predScoreSample}
\title{Example Predicted Pathogenicity Scores}
\format{
A data table with 3848 rows and 13 columns.
\describe{
  \item{sample_name}{\code{character} Sample identifier}
  \item{group}{\code{character} Group classification}
  \item{CHROM}{\code{character} Chromosome where the variant is located}
  \item{POS}{\code{integer} Position of the variant on the chromosome}
  \item{REF}{\code{character} Reference allele}
  \item{ALT}{\code{character} Alternate allele representing the variant}
  \item{genome}{\code{character} Genome build (e.g., hg38)}
  \item{uniprot_id}{\code{character} UniProt identifier for the associated protein}
  \item{transcript_id}{\code{character} Transcript identifier}
  \item{protein_variant}{\code{character} Protein variant notation}
  \item{am_pathogenicity}{\code{numeric} AlphaMissense pathogenicity score}
  \item{am_class}{\code{character} Pathogenicity classification}
  \item{QUAL}{\code{numeric} Quality score of the variant call}
}
}
\source{
Generated using `predictPathoScore` function from the MissensePathoR package.
}
\usage{
predScoreSample
}
\description{
This dataset contains a sample of predicted pathogenicity scores obtained from the
`predictPathoScore` function. It includes a subset of variants with their respective
pathogenicity predictions and classifications.
}
\examples{
data(predScoreSample)
head(predScoreSample)
}
\keyword{datasets}
