% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlphaMissenseData.R
\name{readAlphaMissenseData}
\alias{readAlphaMissenseData}
\title{Read AlphaMissense Data}
\usage{
readAlphaMissenseData(path = NULL)
}
\arguments{
\item{path}{The path to the dataset file.
If NULL, the function loads a demo dataset included in the package.}
}
\value{
A data.table containing the AlphaMissense data.
}
\description{
Retrieves AlphaMissense data for analysis. If no path is provided,
a demo prediction dataset from AlphaMissense is loaded. For the full dataset,
please download it from the provided URL and specify the path to the dataset
in the function.
}
\examples{
\dontrun{
 # Load the demo dataset
 demo_data <- readAlphaMissenseData()

 # Load the full dataset (specify the correct path to the downloaded file)
 # full_data <- readAlphaMissenseData("path/to/AlphaMissense_hg38.tsv.gz")
}
}
