% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF.R
\name{readVCF}
\alias{readVCF}
\title{Read VCF Files into a Data Table}
\usage{
readVCF(vcf_df)
}
\arguments{
\item{vcf_df}{DataFrame with columns 'group', 'sample', 'file_name'.}
}
\value{
A data.table combining all VCF data.
}
\description{
This function processes a set of VCF files described in a DataFrame and combines
them into a single data.table. Each VCF file is associated with a group and sample name.
}
\examples{
vcf_files_path <- system.file("extdata", package = "MissensePathoR")
vcf_df <- createVCFDataFrame(vcf_files_path,
                             "*.vcf",
                             "([0-9]+)h",
                             "Rep([0-9]+)")
vcf_data <- readVCF(vcf_df)

}
