% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMissensePathoR.R
\name{runMissensePathoR}
\alias{runMissensePathoR}
\title{Launch Shiny App for MissensePathoR}
\usage{
runMissensePathoR()
}
\value{
No return value, but opens a Shiny page.
}
\description{
A function that launches the Shiny app for MissensePathoR. This app provides
a user interface to interactively work with the MissensePathoR package.
The app's code is located in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
MissensePathoR::runMissensePathoR()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
