% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSummary.R
\name{scoreSummary}
\alias{scoreSummary}
\title{Calculate Summary Pathogenicity Scores by Specified Category}
\usage{
scoreSummary(data, category)
}
\arguments{
\item{data}{A data table containing pathogenicity scores and categories.}

\item{category}{A character string specifying the column name to group by.}
}
\value{
A data table with mean, median, minimum, maximum, and standard deviation
of the pathogenicity scores for each level of the specified category.
}
\description{
This function calculates various summary statistics of the pathogenicity scores
within the specified category of the data.
}
\examples{
\dontrun{
  scoreSummary(predScoreSample, category = "group")
  scoreSummary(predScoreSample, category = "sample_name")
}
}
