% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreVis.R
\name{scoreVis}
\alias{scoreVis}
\title{Visualize Distribution of Pathogenicity Scores}
\usage{
scoreVis(data, groups = NULL)
}
\arguments{
\item{data}{A data table containing 'am_pathogenicity' and 'group' columns.}

\item{groups}{A character vector specifying the groups to include in the plot.
If NULL, all groups in 'data' are used.}
}
\value{
A ggplot object representing the box plots of pathogenicity score distribution.
}
\description{
This function creates box plots to visualize the distribution of pathogenicity scores within
and between specified groups.
}
\examples{
\dontrun{
  scoreVis(predScoreSample, groups = c("0h", "1h"))
}
}
