% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vcfSample}
\alias{vcfSample}
\title{VCF Sample Data from the MissensePathoR Package}
\format{
A data table with columns:
\describe{
  \item{CHROM}{Chromosome number where the variant is located.}
  \item{POS}{The position of the variant on the chromosome.}
  \item{ID}{The identifier of the variant, if available.}
  \item{REF}{The reference allele at the position.}
  \item{ALT}{The alternate allele observed in the sample.}
  \item{QUAL}{Quality score for the variant calling.}
  \item{FILTER}{Filter status of the variant calling.}
  \item{group}{The experimental group, here indicating the time point after H2O2 introduction.}
  \item{sample_name}{The name of sample, here is the replicate from the experiment.}
}
}
\source{
Generated from VCF files processed using the MissensePathoR package.
From GSE113171 from Gene Expression Omnibus
}
\usage{
vcfSample
}
\description{
This dataset contains combined VCF data for Hela cell replicates across four time points
(0, 1, 4, and 8 hours) after introducing H2O2, processed with the `readVCF` function from
the MissensePathoR package. Chromosomes 1-5 are included, representing a subset of the
human genome.
}
\examples{
\dontrun{
 data(vcfSample)
 head(vcfSample)
}
}
\references{
Rendleman J, Cheng Z, Maity S, et al. New insights into the cellular
temporal response to proteostatic stress. Elife. 2018;7:e39054.
doi: 10.7554/eLife.39054.
}
\keyword{datasets}
